Refactor utilites for HandlePackageFileConflicts
Move statics into helper classes to better seperate responsibility of
types.

Preparing for future refactoring of HandlePackageFileConflicts.
