+++ netstandard/tools/FileUtilities.net45.cs
@@ -2,13 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Build.Utilities;
 using System;
 using System.Reflection;
 
 namespace Microsoft.DotNet.Build.Tasks
 {
-    public partial class HandlePackageFileConflicts : Task
+    static partial class FileUtilities
     {
         private static Version GetAssemblyVersion(string sourcePath)
         {
